Moved all SQL conversions to a separate file
