+++ experiments/AlgebraicRelations/src/AlgebraicRelations.jl
@@ -1,5 +1,6 @@
-module Schema
+module AlgebraicRelations
   include("Presentation.jl")
   include("Query.jl")
+  include("SQL.jl")
   include("Interface.jl")
 end
