+++ experiments/AlgebraicRelations/src/Schema.jl
@@ -1,0 +1,8 @@
+module Presentation
+export Schema
+
+struct Schema{T, U}
+  types::Vector{T}
+  relations::Vector{U}
+end
+end
